Update landing page according to remarks
I tried to update the landing page according to #150 as much as
possible.

I was not able move the css into the HomePage.js. I think I define the
styles for the withStyles method wrong.

Bug #181 is causing problems. Since I changed the redirects, the first
page one sees is always the homepage.