Add script to set up test data in development
This should hopefully address issue #24

The new generate-test-data.js script will create a new organization with a single user profile, and add 10 randomly generated products and 5 randomly generated boxes to it.

It can be run with yarn add-test-data which will prompt the user for the email to use, a password, and the name of the organization to create. After that it will create the data in firestore, and they can log in with the newly created email/password and test things out in the app.