Add react-select to search select fields
I tried to solve #93. It turned out to be quite a hard issue.
Material-UI does not support searches in its Select-Fields itself. They
rather propose solutions using other libraries like react-select
(https://material-ui.com/demos/autocomplete/).
So, I tried to bind a `Select` component to a `TextField` component from Material-UI which in turn
is bound to a Formik `Field`.
I am stuck at the point to forward an onClick event upwards. Each of the
libraries handle changes differently. It is quite a nightmare.