Firestore UI has changed, updating docs to reflect
The workflow in the Google Cloud console for Firestore has changed a little, so updating the instructions.

I also had to update the firebase tooling versions to run the 'yarn run deploy-firestore' command, but I'll submit that separately.

HTTP Error: 410, This version of the Firebase CLI is no longer able to deploy to Firebase. Please upgrade to a newer version (>= 4.1.0). If you have further questions, please reach out to Firebase support.